requirements.txt
requests
streamlit
plotly_express
pandas
numpy
